feat: fetch `metadata_properties` from Argilla
